refactor: home changed to layout, posts as page
