+++ front-vue/layouts/default.vue
@@ -1,4 +1,5 @@
-<script setup>
+<script setup lang="ts">
+import { useToast } from "primevue/usetoast";
 const notificationCount = ref(3);
 
 const menu = ref();
@@ -25,9 +26,26 @@ const items = ref([
   }
 ]);
 
-const toggle = (event) => {
+const toggle = (event: any) => {
   menu.value.toggle(event);
 };
+
+const toast = useToast();
+const handleToast = (status: string, isBookmarked: boolean) => {
+    if (status == 'error') {
+        toast.add({ severity: 'error', summary: 'Error occured', detail: 'Unable to bookmark this post.', life: 3000 });
+        return;
+    }
+
+    if (isBookmarked) {
+        toast.add({ severity: 'info', summary: 'Added to Bookmarks', detail: 'Post added to bookmarks.', life: 3000 });
+    }
+    else if (!isBookmarked) {
+        toast.add({ severity: 'info', summary: 'Removed from Bookmarks', detail: 'Post removed from bookmarks', life: 3000 });
+    }
+
+}
+
 
 </script>
 
@@ -35,13 +53,13 @@ const toggle = (event) => {
   <!-- Sticky navbar with background -->
   <div class="sticky top-0 bg-zinc-900 p-3">
     <div class="flex justify-between items-center">
-      <NuxtLink to="/home">
+      <NuxtLink to="/posts">
         <img src="../assets/img/logov3.svg" alt="Logo" width="40" />
       </NuxtLink>
 
       <div>
         <IconField>
-          <InputText v-model="value1" placeholder="Search" />
+          <InputText placeholder="Search" />
           <InputIcon class="pi pi-search" />
         </IconField>
       </div>
@@ -68,7 +86,36 @@ const toggle = (event) => {
 
   <!-- Content section -->
   <div class="overflow-y-auto">
-    <slot />
+    <div class="flex w-5/6 mx-auto gap-5 py-5">
+        <div class="w-1/5">
+            <PostFilterPanel />
+        </div>
+        <div class="w-3/5">
+          <slot @bookmark-click="handleToast" />
+
+        </div>
+        <div class="w-1/5 flex flex-wrap gap-5">
+            <Card>
+                <template #title>Chats</template>
+                <template #content>
+                    <p class="m-0">
+                        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Inventore sed consequuntur error
+                        repudiandae numquam deserunt quisquam repellat libero asperiores earum nam nobis, culpa ratione
+                        quam perferendis esse, cupiditate neque quas!
+                    </p>
+                </template>
+            </Card>
+
+            <Card class="w-full">
+                <template #title>Contacts</template>
+                <template #content>
+                    <ContactTile />
+                </template>
+            </Card>
+
+
+        </div>
+    </div>
   </div>
