+++ front-vue/pages/register.vue
@@ -14,7 +14,7 @@ const handleRegister = async()=>{
   try {
     console.log(email.value, " ", password.value);
     await register(email.value,username.value, password.value);
-    navigateTo("/home");
+    navigateTo("/posts");
   } catch (error) {
     navigateTo("/");
   }
